GN-4323: Introduce pagination for Fragments
