+++ addon/components/pagination/pagination-view.hbs
@@ -1,0 +1,39 @@
+<div class='au-u-background-gray-100'>
+  <AuToolbar @border='top' @size='large' @nowrap={{true}}>
+    <div class='au-c-pagination'>
+      <p>
+        <span class='au-u-hidden-visually'>
+          {{t 'pagination.results'}}</span>
+        <strong>
+          {{@rangeStart}}
+          -
+          {{@rangeEnd}}
+        </strong>
+        {{t 'pagination.of'}}
+        {{@totalCount}}
+      </p>
+      <div class='au-u-flex'>
+        {{#unless @isFirstPage}}
+          <AuButton
+            @skin='link'
+            @icon='nav-left'
+            @iconAlignment='left'
+            {{on 'click' @onPreviousPage}}
+          >
+            {{t 'pagination.previous'}}
+          </AuButton>
+        {{/unless}}
+        {{#unless @isLastPage}}
+          <AuButton
+            @skin='link'
+            @icon='nav-right'
+            @iconAlignment='right'
+            {{on 'click' @onNextPage}}
+          >
+            {{t 'pagination.next'}}
+          </AuButton>
+        {{/unless}}
+      </div>
+    </div>
+  </AuToolbar>
+</div>
