+++ app/components/pagination/pagination-component.js
@@ -1,0 +1,1 @@
+export { default } from '@lblod/ember-rdfa-editor-lblod-plugins/components/pagination/pagination-component';
